Go binaries are highly compressible. This greatly speeds transfers on slow networks.