@maruel 2017-04-19
The common case for `push.sh` is for both to be on the same local network so in general bandwidth is not too much of a problem. On the other hand single core systems like BeagleBone and CHIP may be significantly slowed down by the compression.

Here's a few knobs that can impact performance:

- with/without `Compression yes` in `~/.ssh/config`
- with/without `rsync --compress`
- hot/warm/cold push
- RPi3 over ethernet (fast case) vs CHIP/BeagleBone over wifi (slow case, still pretty fast)

For cold cache:

- BeagleBone over ethernet, I saw a 1.2s->800ms reduction with `--compress`
- RPi3 over ethernet, I saw 1.1s->650ms
- CHIP over wifi, I saw 1.9s->1.45s
- CHIP over wifi with `Compression yes` in `.ssh/config`, I get 1.45s, with both 1.4s

For hot cache:
- CHIP, it doesn't have much effect at ~1.25s

My main concern was having it to be slower but it doesn't seem to be the case so lgtm.

Note that I have a vague desire to convert this script to Go so that Windows users could leverage it too but that's outside the scope of this change.

```
#!/bin/bash
set -eu
NAME=periph-smoketest.out
go build -o $NAME ./periph-smoketest
HOST=$1
for i in {0..5}; do
  time rsync -v --compress "$NAME" "$HOST:${NAME}"
  # Comment out for hot; we'd need to rebuild with a difference to test for warm
  ssh $HOST "rm $NAME" > /dev/null
done
rm $NAME
```