@jpillora 2017-04-19
Cheers! In general, I find network IO vastly outweighs gzip IO. The only time I'd avoid it would be for files already compressed.