@deadprogram 2017-04-19
I have intentions of adding this capability to Gort, so I am interested in this code. :) 

My own previous attempts at this use scp instead of rsync, not quite as slick:
https://github.com/hybridgroup/gobot-workshop-2017/blob/master/airlock/edison/runner.sh
and
https://github.com/hybridgroup/gobot-workshop-2017/blob/master/airlock/edison/runner.cmd

On related note, part of the intention for adding to Gort is to allow Windows devs to do the same thing. So we're all on the same wavelength.