@maruel 2017-04-19
[OT]
I think [runner.cmd](https://github.com/hybridgroup/gobot-workshop-2017/blob/master/airlock/edison/runner.cmd) has a bug: it doesn't specify [setlocal](https://technet.microsoft.com/library/cc772046.aspx) so the variables leak through after the process run. I.e. "GOARCH and GOOS" remain set after running the script. I don't have access to a Windows box right now so I can't confirm.