@deadprogram 2017-04-19
I doubt anyone tested that, so you are very likely correct.